+++ engine/apps/api/views/shift_swap.py
@@ -42,6 +42,7 @@ class BaseShiftSwapViewSet(ModelViewSet):
 
     def _do_take(self, benefactor: User) -> dict:
         shift_swap = self.get_object()
+        prev_state = shift_swap.insight_logs_serialized
 
         try:
             shift_swap.take(benefactor)
@@ -50,7 +51,13 @@ class BaseShiftSwapViewSet(ModelViewSet):
         except exceptions.BeneficiaryCannotTakeOwnShiftSwapRequest:
             raise BadRequest(detail="A shift swap request cannot be created and taken by the same user")
 
-        update_shift_swap_request_message.apply_async((shift_swap.pk,))
+        write_resource_insight_log(
+            instance=shift_swap,
+            author=self.request.user,
+            event=EntityEvent.UPDATED,
+            prev_state=prev_state,
+            new_state=shift_swap.insight_logs_serialized,
+        )
 
         return ShiftSwapRequestSerializer(shift_swap).data
 
@@ -69,8 +76,6 @@ class BaseShiftSwapViewSet(ModelViewSet):
         return self.serializer_class.setup_eager_loading(queryset)
 
     def perform_destroy(self, instance: ShiftSwapRequest) -> None:
-        # TODO: should we allow deleting a taken request?
-
         super().perform_destroy(instance)
         write_resource_insight_log(instance=instance, author=self.request.user, event=EntityEvent.DELETED)