+++ engine/apps/public_api/tests/test_shift_swap.py
@@ -154,7 +154,7 @@ def test_create_requires_beneficiary(
     make_organization_and_user_with_token,
     make_schedule,
 ):
-    organization, user, token = make_organization_and_user_with_token()
+    organization, _, token = make_organization_and_user_with_token()
 
     schedule = make_schedule(organization, schedule_class=OnCallScheduleWeb)
     today = timezone.now().replace(hour=0, minute=0, second=0, microsecond=0)
@@ -330,15 +330,9 @@ def test_delete(
     mock_update_shift_swap_request_message.apply_async.assert_called_once_with((swap.pk,))
 
 
-@patch("apps.api.views.shift_swap.update_shift_swap_request_message")
 @pytest.mark.django_db
-def test_take(
-    mock_update_shift_swap_request_message,
-    make_organization_and_user_with_token,
-    make_user_for_organization,
-    setup_swap,
-):
-    organization, user, token = make_organization_and_user_with_token()
+def test_take(make_organization_and_user_with_token, make_user_for_organization, setup_swap):
+    organization, _, token = make_organization_and_user_with_token()
     another_user = make_user_for_organization(organization)
     swap = setup_swap(organization)
 
@@ -354,17 +348,10 @@ def test_take(
     assert swap.status == ShiftSwapRequest.Statuses.TAKEN
     assert swap.benefactor == another_user
 
-    mock_update_shift_swap_request_message.apply_async.assert_called_once_with((swap.pk,))
 
-
-@patch("apps.api.views.shift_swap.update_shift_swap_request_message")
 @pytest.mark.django_db
-def test_take_requires_benefactor(
-    mock_update_shift_swap_request_message,
-    make_organization_and_user_with_token,
-    setup_swap,
-):
-    organization, user, token = make_organization_and_user_with_token()
+def test_take_requires_benefactor(make_organization_and_user_with_token, setup_swap):
+    organization, _, token = make_organization_and_user_with_token()
     swap = setup_swap(organization)
 
     client = APIClient()
@@ -378,18 +365,10 @@ def test_take_requires_benefactor(
     assert swap.status == ShiftSwapRequest.Statuses.OPEN
     assert swap.benefactor is None
 
-    mock_update_shift_swap_request_message.apply_async.assert_not_called()
 
-
-@patch("apps.api.views.shift_swap.update_shift_swap_request_message")
 @pytest.mark.django_db
-def test_take_errors(
-    mock_update_shift_swap_request_message,
-    make_organization_and_user_with_token,
-    make_user_for_organization,
-    setup_swap,
-):
-    organization, user, token = make_organization_and_user_with_token()
+def test_take_errors(make_organization_and_user_with_token, make_user_for_organization, setup_swap):
+    organization, _, token = make_organization_and_user_with_token()
     another_user = make_user_for_organization(organization)
     swap = setup_swap(organization)
 
@@ -421,5 +400,3 @@ def test_take_errors(
     data = {"benefactor": another_user.public_primary_key}
     response = client.post(url, data, format="json", HTTP_AUTHORIZATION=f"{token}")
     assert response.status_code == status.HTTP_400_BAD_REQUEST
-
-    mock_update_shift_swap_request_message.apply_async.assert_not_called()